refactor: move number tests
